check distances to neighbor backups while cleaning
This makes algorithms a bit more fair. Imaging backups like this:
1)  19990101-010000
2)  19990101-015000
3)  19990101-016000
4)  19990102-010000

In this situation, original algorithm will choose the folder N3. But
this is one of the most recent backups. And the huge gap between two most
recent backups is going to become even bigger.

Originally, when the script finds 2 commits with smallest logarithmic
distance, it selects the latest one as a victim. Instead, lets check the
logarithmic distance to the neighbors of those 2 commits.

In the example above, candidates for deletion are N2 and N3. Now the
distance N2-N1 is much smaller then N4-N3. So it would be reasonable to
choose directory N2.

Signed-off-by: Yuri Volchkov <yuri.volchkov@gmail.com>