rewrite argument parsing
The interface to this tool became quite unintuitive. This commit is
almost full rework of the argparse.

Alse the option --target-freespace has been disabled for now, because it
makes little sense in its current implementation. It checks if
condition is met right after removing a snapshot. However, the space
freed from the removed snapshot does not appeared in the <df> output
right away. Btrfs has to do some background work to make it happen.

Just disable option for now. Lets figure out later, whether to remove
this functionality, or implement extremely slow version, which waits for
background process to be finished after killing every snapshot.

Signed-off-by: Yuri Volchkov <yuri.volchkov@gmail.com>